base/rag/indexers/apispec/requirements.txt
pymilvus>=2.5.0
httpx>=0.28.0
pyyaml>=6.0
pydantic>=2.0
